Introduce eslint again…
This may be tuned from time to time whenever we find new rules that
makes sense in this project. One thing to notice, if e.g. a callback
function receives three arguments but I only use one or two, prepend the
others with `_` to mark them as there but unused.

Somewhat like this:

```
const middleware = (req, res, _next) => {
  const { body } = req;
  return res.send(body);
};
```

Related: #4