Remove all account templates and create an index instead
Hopefully we don't need these anymore. The index.pug is meant to be the
only required template. It holds the frame and includes the redux
application. So far it's included cross site. Would be nice to have
it from same origin soon.
Besides I commented some JS CDN includes. I'm in a train right now and
maybe they're obsolete anyway.